[#62808308] [#63501138]
[#62808308] The code for co-authors is over, networks is created as a
normal graph rather than a multigraph, with edge attribs as a list for
each user input. Only issue pending is TypeError thrown while trying to
visualize it in gexf format ( not related to this story though). Once
that is over I will 'finish' it in PT and merge it to 'python' repo.

[#63501138] Test cases for the #62808308. Its over. Just need to add 2
more cases and align the code according to 80 chars length.


Former-commit-id: d050e17a2f702e8b1d958b184378e520b9dd5000