[#62809724] [#63120498] [#62820502]
[#62809724] [#63120498]- Now the TypeError will not be raised and it is
handled in wos.py and the error raised because of MAP won't occur
again.
Though the fix is ready for these 2 stories, working on the legible way
of throwing error when the input data is not in the correct format.

[#62820502] The remaining pylint errors are in most of the unused
scripts like bib.py and so on. Except those, most of the genuine cases
were cleared.


Former-commit-id: 454d2165189d97a74719152f31ab5330848b70a3 [formerly f406802b01f203929a1fa40603da7a5226f447d1]
Former-commit-id: 1a75e26417ac72e9b9ceaae51bc6747f63b39067